guiV2_1-main_2/requirements.txt
loguru==0.7.2
matplotlib==3.8.4
numpy==1.26.4
PyYAML==6.0.1
streamlit==1.34.0
tqdm==4.66.4
